Add tabs to organize fields in the weapon sheet
This change adds 3 tabs in the weapon sheet:
 * attributes
 * artifact (for artifacts only)
 * supply and craft (for regular items)

Those tabs share the same styles used in the character sheet.
The supply and artifact tabs are displayed according to the current
artifact die value.
The supply tab, as well as its content will also honor the global
system configuration (showCostField, showSupplyField, etc).

This commit adds new strings for the tab labels. Localization is
available for:
 * en
 * fr
 * sv (courtesy of Petter NystÃm)